# internal/handlers/wallet.go
package handlers

import (
	"net/http"
	"whotterre/argent/internal/dto"
	"whotterre/argent/internal/services"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
)

type WalletHandler struct {
	walletService services.WalletService
}

func NewWalletHandler(walletService services.WalletService) *WalletHandler {
	return &WalletHandler{
		walletService: walletService,
	}
}

// DepositWallet godoc
// @Summary Deposit money into wallet
// @Description Initiate a deposit transaction using Paystack
// @Tags wallet
// @Accept json
// @Produce json
// @Param request body dto.DepositWalletRequest true "Deposit request"
// @Success 200 {object} dto.DepositWalletResponse
// @Failure 400 {object} map[string]string "error"
// @Failure 500 {object} map[string]string "error"
// @Security BearerAuth
// @Router /wallet/deposit [post]
func (h *WalletHandler) Deposit(c *gin.Context) {
	var req dto.DepositWalletRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid input"})
		return
	}

	userID := c.MustGet("user_id").(uuid.UUID)

	response, err := h.walletService.DepositWallet(req, userID)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, response)
}

// GetBalance godoc
// @Summary Get wallet balance
// @Description Retrieve the current balance of the user's wallet
// @Tags wallet
// @Accept json
// @Produce json
// @Success 200 {object} dto.BalanceResponse
// @Failure 500 {object} map[string]string "error"
// @Security BearerAuth
// @Router /wallet/balance [get]
func (h *WalletHandler) GetBalance(c *gin.Context) {
	userID := c.MustGet("user_id").(uuid.UUID)

	balance, err := h.walletService.GetBalance(userID)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, dto.BalanceResponse{Balance: balance})
}

// Transfer godoc
// @Summary Transfer money to another wallet
// @Description Transfer money from user's wallet to another wallet by wallet number
// @Tags wallet
// @Accept json
// @Produce json
// @Param request body dto.TransferRequest true "Transfer request"
// @Success 200 {object} dto.TransferResponse
// @Failure 400 {object} map[string]string "error"
// @Failure 500 {object} map[string]string "error"
// @Security BearerAuth
// @Router /wallet/transfer [post]
func (h *WalletHandler) Transfer(c *gin.Context) {
	var req dto.TransferRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid input"})
		return
	}

	userID := c.MustGet("user_id").(uuid.UUID)

	err := h.walletService.Transfer(userID, req.WalletNumber, req.Amount)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, dto.TransferResponse{Status: "success", Message: "Transfer completed"})
}

// GetTransactions godoc
// @Summary Get transaction history
// @Description Retrieve the user's transaction history
// @Tags wallet
// @Accept json
// @Produce json
// @Success 200 {array} dto.TransactionResponse
// @Failure 500 {object} map[string]string "error"
// @Security BearerAuth
// @Router /wallet/transactions [get]
func (h *WalletHandler) GetTransactions(c *gin.Context) {
	userID := c.MustGet("user_id").(uuid.UUID)

	transactions, err := h.walletService.GetTransactions(userID)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	var response []dto.TransactionResponse
	for _, t := range transactions {
		response = append(response, dto.TransactionResponse{
			Type:   t.Type,
			Amount: t.Amount,
			Status: t.Status,
		})
	}

	c.JSON(http.StatusOK, response)
}

// GetDepositStatus godoc
// @Summary Get deposit transaction status
// @Description Check the status of a deposit transaction by reference
// @Tags wallet
// @Accept json
// @Produce json
// @Param reference path string true "Transaction reference"
// @Success 200 {object} map[string]interface{}
// @Failure 404 {object} map[string]string "error"
// @Security BearerAuth
// @Router /wallet/deposit/{reference}/status [get]
func (h *WalletHandler) GetDepositStatus(c *gin.Context) {
	reference := c.Param("reference")

	status, err := h.walletService.GetDepositStatus(reference)
	if err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": "Transaction not found"})
		return
	}

	c.JSON(http.StatusOK, status)
}

// Webhook godoc
// @Summary Process Paystack webhook
// @Description Handle webhook notifications from Paystack for payment events
// @Tags wallet
// @Accept json
// @Produce json
// @Param X-Paystack-Signature header string true "Paystack webhook signature"
// @Success 200 {object} map[string]bool "status"
// @Failure 400 {object} map[string]string "error"
// @Router /wallet/paystack/webhook [post]
func (h *WalletHandler) Webhook(c *gin.Context) {
	payload, _ := c.GetRawData()
	signature := c.GetHeader("x-paystack-signature")

	err := h.walletService.ProcessWebhook(payload, signature)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, gin.H{"status": true})
}

// DepositCallback godoc
// @Summary Handle deposit callback
// @Description Handle callback from Paystack after deposit payment
// @Tags wallet
// @Accept json
// @Produce json
// @Param reference query string true "Transaction reference"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]string "error"
// @Failure 404 {object} map[string]string "error"
// @Router /wallet/deposit/callback [get]
func (h *WalletHandler) DepositCallback(c *gin.Context) {
	reference := c.Query("reference")
	if reference == "" {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Reference missing"})
		return
	}

	status, err := h.walletService.GetDepositStatus(reference)
	if err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": "Transaction not found"})
		return
	}

	c.JSON(http.StatusOK, status)
}